+++ docs/changelog.md
@@ -4,7 +4,8 @@
 ### Features
 * Added a functioning Nimbus CL client
 * Added Nethermind EL client
-* Added a Lodestar CL beacon Launcher
+* Added a Lodestar CL (beacon, validator) Launcher
+* * Added new `GetNodeSyncingData` method in cl rest client
 * Added a transaction spammer to blast the network with transactions after all the nodes come up
 * Added optional waiting until epoch finalization occurs